+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 ### New Features
 
 * Add aws-sso exec --overwrite-env flag to allow overriding AWS_ env variables #1095
+* Add support for /healthcheck endpoint for ECS server #1356
 
 ### Changes